requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0
matplotlib>=3.5.0
plotly>=5.0.0

# Configuration and logging
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0
tensorboard>=2.10.0

# Computer Vision and Video Processing
opencv-python>=4.6.0
kornia>=0.6.0
decord>=0.6.0
av>=10.0.0

# Data handling
Pillow>=9.0.0
scikit-image>=0.19.0
scipy>=1.8.0

# Development and testing
black>=22.0.0
ruff>=0.1.0
pytest>=7.0.0
pytest-cov>=4.0.0

# Demo and visualization
streamlit>=1.25.0
gradio>=3.40.0

# Optional: Advanced features
# detectron2 @ git+https://github.com/facebookresearch/detectron2.git
# pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git
